+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 <a name="Pending Release"></a>
 ## [Pending Release](https://github.com/lightstep/otel-launcher-java/compare/1.5.0...master)
+* Use OpenTelemetry Java API/SDK version 1.12.0
+* Use OpenTelemetry Java Instrumentation version 1.12.1
+* Removed the usage of the oshi instrumentation (used to report host metrics).
 
 <a name="1.6.0"></a>
 ## [1.6.0](https://github.com/lightstep/otel-launcher-java/compare/1.4.0...1.5.0)